requirements.txt
google-adk
groq==0.9.0
pydantic==2.6.4
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.3
trafilatura==1.8.0
jinja2==3.1.3
python-dateutil==2.8.2
tenacity==8.2.3
loguru==0.7.2
httpx==0.27.0
markdown==3.5.2
python-dotenv==1.0.1
google-generativeai==0.5.4
python-dotenv